+++ src/components/ReportsCategoriesNav/ReportsCategoriesNavStyles.module.css
@@ -8,6 +8,7 @@
 .reportsCategoriesNavNextButton {
   border: none;
   cursor: pointer;
+  background: none;
   padding: 0;
   display: flex;
   justify-content: center;
